# loop_implementation/api/openapi.yaml
openapi: 3.1.0
info:
  title: Loop Platform API
  version: 0.1.0
  description: |
    Public REST API for the Loop agent platform.

    All endpoints require a Bearer token scoped to a workspace.
    Streaming endpoints additionally support Server-Sent Events (SSE) and WebSocket.
  contact:
    name: Loop Engineering
    email: eng@loop.example
  license:
    name: Apache-2.0
servers:
  - url: https://api.loop.example/v1
    description: Production (na-east — abstract region; concrete cloud may be AWS/Azure/GCP/Alibaba/self-host)
  - url: https://api-eu.loop.example/v1
    description: Production (eu-west — same)
  - url: http://localhost:8080/v1
    description: Local dev
security:
  - BearerAuth: []
...
paths:
...
  /workspaces:
    get:
      tags: [Workspaces]
      summary: List workspaces the caller is a member of
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceList"
    post:
      tags: [Workspaces]
      summary: Create workspace
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkspaceCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
  /workspaces/{workspace_id}:
    get:
      tags: [Workspaces]
      summary: Get workspace
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/RequestRegion" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Workspaces]
      summary: Update workspace metadata
      description: Workspace `region` is immutable after creation.
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/RequestRegion" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkspacePatch" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /workspaces/{workspace_id}/inbox:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [Inbox]
      summary: List pending and caller-claimed inbox items
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InboxItemList" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /workspaces/{workspace_id}/inbox/escalate:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    post:
      tags: [Inbox]
      summary: Escalate a conversation into the operator inbox
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InboxEscalateRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InboxItem" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Conversation already has an open inbox item
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemDetail" }
...
  /inbox/{item_id}/claim:
    parameters: [{ $ref: "#/components/parameters/InboxItemId" }]
    post:
      tags: [Inbox]
      summary: Claim an inbox item for the caller or named operator
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InboxClaimRequest" }
      responses:
        "200":
          description: Claimed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InboxItem" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Item is no longer claimable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemDetail" }
  /inbox/{item_id}/release:
    parameters: [{ $ref: "#/components/parameters/InboxItemId" }]
    post:
      tags: [Inbox]
      summary: Release a claimed inbox item back to pending
      responses:
        "200":
          description: Released
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InboxItem" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Item is not currently claimed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemDetail" }
  /inbox/{item_id}/resolve:
    parameters: [{ $ref: "#/components/parameters/InboxItemId" }]
    post:
      tags: [Inbox]
      summary: Resolve a claimed inbox item
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InboxResolveRequest" }
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InboxItem" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Item is not currently claimed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemDetail" }
...
  /agents:
    post:
      tags: [Agents]
      summary: Create an agent
      parameters:
        - { $ref: "#/components/parameters/ActiveWorkspaceId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
    get:
      tags: [Agents]
      summary: List agents in the active workspace
      parameters:
        - { $ref: "#/components/parameters/ActiveWorkspaceId" }
      responses:
        "200":
          description: OK
          headers:
            X-RateLimit-Limit:
              schema: { type: integer }
            X-RateLimit-Remaining:
              schema: { type: integer }
            X-RateLimit-Reset:
              schema: { type: integer }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /traces/{turn_id}:
    parameters: [{ $ref: "#/components/parameters/TurnId" }]
    get:
      tags: [Traces]
      summary: Full trace detail for a turn
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Trace" }
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PASETO
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema: { type: integer, description: "Seconds to wait before retrying" }
        X-RateLimit-Limit:
          schema: { type: integer }
        X-RateLimit-Remaining:
          schema: { type: integer }
        X-RateLimit-Reset:
          schema: { type: integer, description: "Unix timestamp of reset" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
...
  parameters:
    WorkspaceId:
      in: path
      name: workspace_id
      required: true
      schema: { type: string, format: uuid }
    ActiveWorkspaceId:
      in: header
      name: X-Loop-Workspace-Id
      required: true
      schema: { type: string, format: uuid }
      description: Workspace selected for active agent and audit routes.
    RequestRegion:
      in: header
      name: X-Loop-Region
      required: false
      description: Abstract region serving the request; mismatches with workspace.region return 403.
      schema: { $ref: "#/components/schemas/RegionName" }
...
    InboxItemId:
      in: path
      name: item_id
      required: true
      schema: { type: string, format: uuid }
...
    TurnId:
      in: path
      name: turn_id
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    ProblemDetail:
      description: RFC 9457 Problem Details for HTTP APIs
      type: object
      required: [type, status, title]
      properties:
        type: { type: string, format: uri, example: "https://api.loop.example/errors/rate-limit" }
        status: { type: integer, example: 429 }
        title: { type: string, example: "Too Many Requests" }
        detail: { type: string }
        instance: { type: string, format: uri }
        request_id: { type: string, description: "Correlation ID for logs" }
    Workspace:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        plan: { type: string, enum: [hobby, pro, team, enterprise] }
        region: { $ref: "#/components/schemas/RegionName" }
        created_at: { type: string, format: date-time }
        role: { type: string, enum: [owner, admin, member, viewer] }
    WorkspaceList:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Workspace" }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
    WorkspaceCreate:
      type: object
      required: [name, slug]
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        slug: { type: string, pattern: "^[a-z0-9][a-z0-9-]*$" }
        region:
          allOf: [{ $ref: "#/components/schemas/RegionName" }]
          default: na-east
          description: Immutable workspace region pin selected at creation.
    WorkspacePatch:
      type: object
      description: Mutable workspace metadata. Region changes are rejected.
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
    RegionName:
      type: string
      enum: [na-east, eu-west]
      description: Abstract Loop region from infra/terraform/regions.yaml.
    Agent:
      type: object
      properties:
        id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        active_version: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
    AgentList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Agent" }
    AgentCreate:
      type: object
      required: [name, slug]
      properties:
        name: { type: string }
        slug: { type: string, pattern: "^[a-z0-9-]+$" }
        description: { type: string }
...
    InboxStatus:
      type: string
      enum: [pending, claimed, resolved]
    InboxChannel:
      type: string
      enum: [web, voice, sms, whatsapp, slack]
    InboxItem:
      type: object
      required:
        - id
        - workspace_id
        - agent_id
        - conversation_id
        - user_id
        - status
        - reason
        - operator_id
        - created_at_ms
        - claimed_at_ms
        - resolved_at_ms
        - last_message_excerpt
      properties:
        id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        agent_id: { type: string, format: uuid }
        conversation_id: { type: string, format: uuid }
        user_id: { type: string }
        status: { $ref: "#/components/schemas/InboxStatus" }
        reason: { type: string }
        operator_id: { type: string, nullable: true }
        created_at_ms: { type: integer }
        claimed_at_ms: { type: integer, nullable: true }
        resolved_at_ms: { type: integer, nullable: true }
        last_message_excerpt: { type: string }
        team_id:
          type: string
          description: Optional Studio grouping metadata when an inbox backend provides it.
        channel:
          $ref: "#/components/schemas/InboxChannel"
    InboxItemList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/InboxItem" }
    InboxEscalateRequest:
      type: object
      required: [agent_id, conversation_id, user_id, reason]
      properties:
        agent_id: { type: string, format: uuid }
        conversation_id: { type: string, format: uuid }
        user_id: { type: string, minLength: 1 }
        reason: { type: string, minLength: 1 }
        last_message_excerpt: { type: string }
        now_ms: { type: integer, minimum: 0 }
    InboxClaimRequest:
      type: object
      properties:
        operator_id: { type: string, nullable: true }
        now_ms: { type: integer, minimum: 0 }
    InboxResolveRequest:
      type: object
      properties:
        now_ms: { type: integer, minimum: 0 }
...
    Trace:
      type: object
      properties:
        turn_id: { type: string, format: uuid }
        spans:
          type: array
          items:
            type: object
            properties:
              span_id: { type: string }
              parent_span_id: { type: string, nullable: true }
              kind: { type: string, enum: [llm, tool, retrieval, memory, channel] }
              name: { type: string }
              started_at: { type: string, format: date-time }
              latency_ms: { type: integer }
              cost_usd: { type: number }
              status: { type: string }
              attrs: { type: object }
...